Refactor config cmd
